Move migration code to their own files
